Add: separate builders into their own structs to reduce QueryBuilder noise and added a macro to automatically space clauses
