--- src/states.rs
@@ -1,3 +1,0 @@
-pub struct DefaultState;
-pub struct SelectState;
-pub struct FromReady;
